Fix #1333: merge some test files to do less builds during e2e tests
